Android 4.4.3 release 1

Conflicts:
	android_webview/java/src/org/chromium/android_webview/AwContents.java
	base/android/jni_generator/jni_generator_tests.py

+++ android_webview/java/src/org/chromium/android_webview/AwQuotaManagerBridge.java
@@ -1,17 +1,17 @@
-// Copyright (c) 2013 The Chromium Authors. All rights reserved.
+// Copyright 2013 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
 package org.chromium.android_webview;
 
+import android.webkit.ValueCallback;
+
 import org.chromium.base.CalledByNative;
 import org.chromium.base.JNINamespace;
 import org.chromium.base.ThreadUtils;
 
-import android.webkit.ValueCallback;
-
-import java.util.Map;
 import java.util.HashMap;
+import java.util.Map;
 
 /**
  * Bridge between android.webview.WebStorage and native QuotaManager. This object is owned by Java
@@ -22,7 +22,7 @@ import java.util.HashMap;
 @JNINamespace("android_webview")
 public class AwQuotaManagerBridge {
     // TODO(boliu): This should be obtained from Java AwBrowserContext that owns this.
-    private static native int nativeGetDefaultNativeAwQuotaManagerBridge();
+    private static native long nativeGetDefaultNativeAwQuotaManagerBridge();
 
     // TODO(boliu): This should be owned by Java AwBrowserContext, not a singleton.
     private static AwQuotaManagerBridge sInstance;
@@ -52,7 +52,7 @@ public class AwQuotaManagerBridge {
     }
 
     // This is not owning. The native object is owned by the native AwBrowserContext.
-    private int mNativeAwQuotaManagerBridgeImpl;
+    private long mNativeAwQuotaManagerBridgeImpl;
 
     // The Java callbacks are saved here. An incrementing callback id is generated for each saved
     // callback and is passed to the native side to identify callback.
@@ -61,7 +61,7 @@ public class AwQuotaManagerBridge {
     private Map<Integer, ValueCallback<Long>> mPendingGetQuotaForOriginCallbacks;
     private Map<Integer, ValueCallback<Long>> mPendingGetUsageForOriginCallbacks;
 
-    private AwQuotaManagerBridge(int nativeAwQuotaManagerBridgeImpl) {
+    private AwQuotaManagerBridge(long nativeAwQuotaManagerBridgeImpl) {
         mNativeAwQuotaManagerBridgeImpl = nativeAwQuotaManagerBridgeImpl;
         mPendingGetOriginCallbacks =
                 new HashMap<Integer, ValueCallback<Origins>>();
@@ -145,18 +145,18 @@ public class AwQuotaManagerBridge {
     private void onGetUsageAndQuotaForOriginCallback(
             int callbackId, boolean isQuota, long usage, long quota) {
         if (isQuota) {
-          assert mPendingGetQuotaForOriginCallbacks.containsKey(callbackId);
-          mPendingGetQuotaForOriginCallbacks.remove(callbackId).onReceiveValue(quota);
+            assert mPendingGetQuotaForOriginCallbacks.containsKey(callbackId);
+            mPendingGetQuotaForOriginCallbacks.remove(callbackId).onReceiveValue(quota);
         } else {
-          assert mPendingGetUsageForOriginCallbacks.containsKey(callbackId);
-          mPendingGetUsageForOriginCallbacks.remove(callbackId).onReceiveValue(usage);
+            assert mPendingGetUsageForOriginCallbacks.containsKey(callbackId);
+            mPendingGetUsageForOriginCallbacks.remove(callbackId).onReceiveValue(usage);
         }
     }
 
-    private native void nativeInit(int nativeAwQuotaManagerBridgeImpl);
-    private native void nativeDeleteAllData(int nativeAwQuotaManagerBridgeImpl);
-    private native void nativeDeleteOrigin(int nativeAwQuotaManagerBridgeImpl, String origin);
-    private native void nativeGetOrigins(int nativeAwQuotaManagerBridgeImpl, int callbackId);
-    private native void nativeGetUsageAndQuotaForOrigin(int nativeAwQuotaManagerBridgeImpl,
+    private native void nativeInit(long nativeAwQuotaManagerBridgeImpl);
+    private native void nativeDeleteAllData(long nativeAwQuotaManagerBridgeImpl);
+    private native void nativeDeleteOrigin(long nativeAwQuotaManagerBridgeImpl, String origin);
+    private native void nativeGetOrigins(long nativeAwQuotaManagerBridgeImpl, int callbackId);
+    private native void nativeGetUsageAndQuotaForOrigin(long nativeAwQuotaManagerBridgeImpl,
             String origin, int callbackId, boolean isQuota);
 }
