+++ chrome/android/java/src/org/chromium/chrome/browser/ContentViewUtil.java
@@ -1,4 +1,4 @@
-// Copyright (c) 2012 The Chromium Authors. All rights reserved.
+// Copyright 2012 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -17,17 +17,17 @@ public abstract class ContentViewUtil {
      * @return pointer to native WebContents instance, suitable for using with a
      *         (java) ContentViewCore instance.
      */
-    public static int createNativeWebContents(boolean incognito) {
+    public static long createNativeWebContents(boolean incognito) {
         return nativeCreateNativeWebContents(incognito);
     }
 
     /**
      * @param webContentsPtr The WebContents reference to be deleted.
      */
-    public static void destroyNativeWebContents(int webContentsPtr) {
+    public static void destroyNativeWebContents(long webContentsPtr) {
         nativeDestroyNativeWebContents(webContentsPtr);
     }
 
-    private static native int nativeCreateNativeWebContents(boolean incognito);
-    private static native void nativeDestroyNativeWebContents(int webContentsPtr);
+    private static native long nativeCreateNativeWebContents(boolean incognito);
+    private static native void nativeDestroyNativeWebContents(long webContentsPtr);
 }
